backend/requirements.txt
fastapi
uvicorn
google-cloud-firestore
google-cloud-storage
pydantic
httpx
geopy
pydantic-settings